Force ert to exit with a non 0 errcode + errmsg in case of test fail
As it was, `make test` couldn't exit with a non 0 exit code, even in
case of an error. Github actions are relying on the job exit code to
associate a success code. Meaning, the testsuite couldn't ever fail.

Always green CI is highly desirable, but in that case, the status was
perhaps a bit misleading :P

Note: we have the same issue with the batch-byte-compile function. I
couldn't find an easy fix for that one.