Let info.el parse INFOPATH
A trailing colon in INFOPATH indicates select default directories should
be appended to `Info-directory-list'. See `Info-default-dirs'.

`parse-colon-path' produces a trailing `nil' in `Info-directory-list'
when a trailing colon is present in INFOPATH. This leads to type errors
in Info functions expecting only strings.

`info-initialize' is called by Info entry points and parses INFOPATH
when `Info-directory-list' is not already set. Clear it to force parsing
instead of handling parsing ourselves.

Fixes: 7effcda0c987 ("Update info-directory-list from INFOPATH if set")
Closes: #25