Modify mode-line lighter customization
This adds a new customizable variable, `envrc-lighter-function` that sits
between `envrc-lighter` and `envrc-STATUS-lighter`. This allows overriding
`envrc-lighter` without relying on internal definitions (`envrc--status`).

The other benefit of this is that any customization is likely to want to set all
three `envrc-STATUS-lighter`, with them being mostly the same, with a small
conditional part (as you can see from both `envrc--default-lighter` and my
lambda below). While this variable is a `risky-local-variable`, I believe that
`envrc-STATUS-lighter` are also risky, because they can also contain `:eval`,
no?

If `envrc-lighter-function` is not customized, the existing
`envrc-STATUS-lighter` variables should continue to work as they have.

In future, I think this could be the _only_ `-lighter` customizable variable.

I like my minor modes to be pretty minimal, so I use this like
```elisp
(setq envrc-lighter-function
      (lambda (status)
        `(:propertize "🗁"
                      face
                      ,(pcase status
                         ('on 'envrc-mode-line-on-face)
                         ('error 'envrc-mode-line-error-face)
                         (_ 'envrc-mode-line-none-face)))))
```